extract GlobalFilter component for table and grouped row table
