+++ packages/data-tables/src/components/GlobalFilter.js
@@ -1,0 +1,29 @@
+import React, { useState } from 'react';
+import PropTypes from 'prop-types';
+import { useAsyncDebounce } from 'react-table';
+
+const GlobalFilter = ({ globalFilter, setGlobalFilter }) => {
+  const [value, setValue] = useState(globalFilter);
+  const onChange = useAsyncDebounce((value) => {
+    setGlobalFilter(value || undefined);
+  }, 200);
+
+  return (
+    <input
+      value={value || ''}
+      onChange={(e) => {
+        setValue(e.target.value);
+        onChange(e.target.value);
+      }}
+      placeholder={`Search all columns...`}
+      type="search"
+    />
+  );
+};
+
+GlobalFilter.propTypes = {
+  globalFilter: PropTypes.string,
+  setGlobalFilter: PropTypes.func,
+};
+
+export default GlobalFilter;
